feat: move component's types to corresponding files
